pdf_table_extractor/requirements.txt
# Core dependencies
openai>=1.12.0

# PDF processing
PyMuPDF>=1.23.0    # Also known as 'fitz' - fast PDF processing, no external dependencies
Pillow>=10.0.0

# Data validation
pydantic>=2.5.0

# Utilities
tqdm>=4.66.0

# Optional: For loading .env files (can also use system environment variables)
python-dotenv>=1.0.0

# Optional: For advanced table detection
# opencv-python>=4.8.0
# numpy>=1.24.0

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=7.4.0
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.5.0
